Struct v8::error::StackFrame

pub struct StackFrame(_, _);

A stack frame, that is bound to an isolate.

Methods

impl StackFrame

fn get_line_number(&self) -> u32

The line number at which this stack frame was pushed.

fn get_column(&self) -> u32

The column number at which this stack frame was pushed.

fn get_script_name(&self) -> Option<String>

The script file name in which this stack frame was pushed.

fn get_function_name(&self) -> String

The function name in which this stack frame was pushed.

fn is_eval(&self) -> bool

Whether this stack frame is part of an eval call.

fn is_constructor(&self) -> bool

Whether this stack frame is part of a constructor call.

fn to_captured(&self) -> CapturedStackFrame

Creates a captured version of this stack frame, that doesn't retain a reference to its isolate.