Module v8::inspector

Bindings to the V8 Inspector API. Documentation for the V8 inspector API is very sparse, so here are a few references for the next sorry soul who has to dig into it.

https://medium.com/@hyperandroid/v8-inspector-from-an-embedder-standpoint-7f9c0472e2b7 https://v8.dev/docs/inspector https://chromedevtools.github.io/debugger-protocol-viewer/tot/ https://cs.chromium.org/chromium/src/v8/include/v8-inspector.h https://github.com/nodejs/node/blob/v13.7.0/src/inspector_agent.cc https://github.com/nodejs/node/blob/v13.7.0/src/inspector_agent.h https://github.com/nodejs/node/tree/v13.7.0/src/inspector https://github.com/denoland/deno/blob/v0.38.0/cli/inspector.rs

Structs

Channel

ChannelBase

CharacterArray

StringBuffer

StringViewIterator

V8Inspector

V8InspectorClient

V8InspectorClientBase

V8InspectorSession

V8StackTrace

Enums

StringView

Traits

AsChannel

AsV8InspectorClient

ChannelImpl

V8InspectorClientImpl

Functions

v8_inspector__V8InspectorClient__BASE__consoleAPIMessage⚠

v8_inspector__V8InspectorClient__BASE__generateUniqueId⚠

v8_inspector__V8InspectorClient__BASE__quitMessageLoopOnPause⚠

v8_inspector__V8InspectorClient__BASE__runIfWaitingForDebugger⚠

v8_inspector__V8InspectorClient__BASE__runMessageLoopOnPause⚠

v8_inspector__V8Inspector__Channel__BASE__flushProtocolNotifications⚠

v8_inspector__V8Inspector__Channel__BASE__sendNotification⚠

v8_inspector__V8Inspector__Channel__BASE__sendResponse⚠