Module v8::script_compiler

Structs

CachedData

Compilation data that the embedder can cache and pass back to speed up future compilations. The data is produced if the CompilerOptions passed to the compilation functions in ScriptCompiler contains produce_data_to_cache = true. The data to cache can then can be retrieved from UnboundScript.

Source

Source code which can then be compiled to a UnboundScript or Script.

Enums

CompileOptions

NoCacheReason

The reason for which we are not requesting or providing a code cache.

Functions

cached_data_version_tag

Return a version tag for CachedData for the current V8 version & flags.

compile

compile_function

compile_module

Compile an ES module, returning a Module that encapsulates the compiled code.

compile_module2

Same as compile_module with more options.

compile_unbound_script