Struct v8::HandleScope

pub struct HandleScope<'s, C = Context> { /* private fields */ }

A stack-allocated class that governs a number of local handles. After a handle scope has been created, all local handles will be allocated within that handle scope until either the handle scope is deleted or another handle scope is created. If there is already a handle scope and a new one is created, all allocations will take place in the new handle scope until it is deleted. After that, new handles will again be allocated in the original handle scope.

After the handle scope of a local handle has been deleted the garbage collector will no longer track the object stored in the handle and may deallocate it. The behavior of accessing a handle for which the handle scope has been deleted is undefined.

Implementations

impl<'s> HandleScope<'s>

pub fn new<P: NewHandleScope<'s>>(param: &'s mut P) -> P::NewScope

pub fn with_context<P: NewHandleScopeWithContext<'s>, H: Handle<Data = Context>>(
param: &'s mut P,
context: H
) -> Self

Opens a new HandleScope and enters a Context in one step. The first argument should be an Isolate or OwnedIsolate. The second argument can be any handle that refers to a Context object; usually this will be a Global<Context>.

pub fn get_current_context(&self) -> Local<'s, Context>

Returns the context of the currently running JavaScript, or the context on the top of the stack if no JavaScript is running.

pub fn get_entered_or_microtask_context(&self) -> Local<'s, Context>

Returns either the last context entered through V8’s C++ API, or the context of the currently running microtask while processing microtasks. If a context is entered while executing a microtask, that context is returned.

impl<'s> HandleScope<'s, ()>

pub fn throw_exception(
&mut self,
exception: Local<'_, Value>
) -> Local<'s, Value>

Schedules an exception to be thrown when returning to JavaScript. When an exception has been scheduled it is illegal to invoke any JavaScript operation; the caller must return immediately and only after the exception has been handled does it become legal to invoke JavaScript operations.

This function always returns the undefined value.

impl<'s> HandleScope<'s>

pub fn get_isolate_data_from_snapshot_once<T>(
&mut self,
index: usize
) -> Result<Local<'s, T>, DataError> where
T: 'static,
for<'l> Local<'l, Data>: TryInto<Local<'l, T>, Error = DataError>,

Return data that was previously attached to the isolate snapshot via SnapshotCreator, and removes the reference to it. If called again with same index argument, this function returns DataError::NoData.

The value that was stored in the snapshot must either match or be convertible to type parameter T, otherwise DataError::BadType is returned.

pub fn get_context_data_from_snapshot_once<T>(
&mut self,
index: usize
) -> Result<Local<'s, T>, DataError> where
T: 'static,
for<'l> Local<'l, Data>: TryInto<Local<'l, T>, Error = DataError>,

Return data that was previously attached to the context snapshot via SnapshotCreator, and removes the reference to it. If called again with same index argument, this function returns DataError::NoData.

The value that was stored in the snapshot must either match or be convertible to type parameter T, otherwise DataError::BadType is returned.