Struct v8::template::ObjectTemplate

pub struct ObjectTemplate<'a>(_, _);

Methods

impl<'a> ObjectTemplate<'a>

An ObjectTemplate is used to create objects at runtime.

Properties added to an ObjectTemplate are added to each object created from the ObjectTemplate.

fn new(isolate: &'a Isolate) -> ObjectTemplate<'a>

fn set_internal_field_count(&self, value: usize)

fn new_instance(&self, context: &Context) -> Object<'a>

unsafe fn from_raw(isolate: &'a Isolate, raw: ObjectTemplateRef) -> ObjectTemplate<'a>

Creates an object template from a set of raw pointers.

fn as_raw(&self) -> ObjectTemplateRef

Returns the underlying raw pointer behind this object template.