Module v8::value

Javascript values that user code can interact with.

Structs

Array
An instance of the built-in array constructor (ECMA-262, 15.4.2).

ArrayBuffer
An instance of the built-in ArrayBuffer constructor (ES6 draft 15.13.5).

ArrayBufferView
A base class for an instance of one of "views" over ArrayBuffer, including TypedArrays and DataView (ES6 draft 15.13).

Boolean
A primitive boolean value (ECMA-262, 4.3.14). Either the true or false value.

BooleanObject
A Boolean object (ECMA-262, 4.3.15).

Data
The superclass of values and API object templates.

DataView
An instance of DataView constructor (ES6 draft 15.13.7).

Date
An instance of the built-in Date constructor (ECMA-262, 15.9).

Exception

External
A JavaScript value that wraps an external value. This type of value is mainly used to associate native data structures with JavaScript objects.

Float32Array
An instance of Float32Array constructor (ES6 draft 15.13.6).

Float64Array
An instance of Float64Array constructor (ES6 draft 15.13.6).

Function
A JavaScript function object (ECMA-262, 15.3).

FunctionCallbackInfo

Int16Array
An instance of Int16Array constructor (ES6 draft 15.13.6).

Int32
A JavaScript value representing a 32-bit signed integer.

Int32Array
An instance of Int32Array constructor (ES6 draft 15.13.6).

Int8Array
An instance of Int8Array constructor (ES6 draft 15.13.6).

Integer
A JavaScript value representing a signed integer.

Map
An instance of the built-in Map constructor (ECMA-262, 6th Edition, 23.1.1).

Name
A superclass for symbols and strings.

Number
A JavaScript number value (ECMA-262, 4.3.20)

NumberObject
A Number object (ECMA-262, 4.3.21).

Object
A JavaScript object (ECMA-262, 4.3.3)

Primitive
The superclass of primitive values. See ECMA-262 4.3.2.

Private
A private symbol

Promise
An instance of the built-in Promise constructor (ES6 draft).

PropertyCallbackInfo

Proxy
An instance of the built-in Proxy constructor (ECMA-262, 6th Edition, 26.2.1).

RegExp
An instance of the built-in RegExp constructor (ECMA-262, 15.10).

Set
An instance of the built-in Set constructor (ECMA-262, 6th Edition, 23.2.1).

SharedArrayBuffer
An instance of the built-in SharedArrayBuffer constructor.

String
A JavaScript string value (ECMA-262, 4.3.17).

StringObject
A String object (ECMA-262, 4.3.18).

Symbol
A JavaScript symbol (ECMA-262 edition 6)

SymbolObject
A Symbol object (ECMA-262 edition 6).

TypedArray
A base class for an instance of TypedArray series of constructors (ES6 draft 15.13.6).

Uint16Array
An instance of Uint16Array constructor (ES6 draft 15.13.6).

Uint32
A JavaScript value representing a 32-bit unsigned integer.

Uint32Array
An instance of Uint32Array constructor (ES6 draft 15.13.6).

Uint8Array
An instance of Uint8Array constructor (ES6 draft 15.13.6).

Uint8ClampedArray
An instance of Uint8ClampedArray constructor (ES6 draft 15.13.6).

Value
The superclass of all JavaScript values and objects.

Functions

false_

null

true_

undefined

Type Definitions

FunctionCallback